+++ operator/cnp_derivative.go
@@ -12,16 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package groups
+package main
 
 import (
 	"fmt"
+	"time"
 
-	"github.com/cilium/cilium/pkg/k8s"
+	"github.com/cilium/cilium/pkg/controller"
 	cilium_v2 "github.com/cilium/cilium/pkg/k8s/apis/cilium.io/v2"
 	clientset "github.com/cilium/cilium/pkg/k8s/client/clientset/versioned"
 	"github.com/cilium/cilium/pkg/lock"
+	"github.com/cilium/cilium/pkg/logging/logfields"
 	"github.com/cilium/cilium/pkg/policy/api"
+	_ "github.com/cilium/cilium/pkg/policy/groups"
+
+	"github.com/sirupsen/logrus"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
@@ -29,9 +34,15 @@ const (
 	cnpKindName = "derivative"
 	parentCNP   = "io.cilium.network.policy.parent.uuid"
 	cnpKindKey  = "io.cilium.network.policy.kind"
+
+	// maxNumberOfAttempts Number of times that try to retrieve a information from a cloud provider.
+	maxNumberOfAttempts = 5
+	// SleepDuration time that sleep in case that can't retrieve information from a cloud provider.
+	sleepDuration = 5 * time.Second
 )
 
 var (
+	controllerManager     = controller.NewManager()
 	globalK8sClient       *clientset.Clientset
 	k8sMutex              = lock.Mutex{}
 	blockOwnerDeletionPtr = true
@@ -98,51 +109,21 @@ func denyEgressRule() *api.Rule {
 	}
 }
 
-// getK8sClient return the kubernetes apiserver connection
-func getK8sClient() (*clientset.Clientset, error) {
-	k8sMutex.Lock()
-	defer k8sMutex.Unlock()
-	if globalK8sClient != nil {
-		return globalK8sClient, nil
-	}
-
-	restConfig, err := k8s.CreateConfig()
-	if err != nil {
-		return nil, fmt.Errorf("Unable to create rest configuration: %s", err)
-	}
-	k8sClient, err := clientset.NewForConfig(restConfig)
-	if err != nil {
-		return nil, fmt.Errorf("Unable to create Kubernetes configuration: %s", err)
-
-	}
-	globalK8sClient = k8sClient
-	return globalK8sClient, nil
-}
-
 func updateCNPStatus(cnp *cilium_v2.CiliumNetworkPolicy) (*cilium_v2.CiliumNetworkPolicy, error) {
-	k8sClient, err := getK8sClient()
-	if err != nil {
-		// @TODO change the error here
-		return nil, err
-	}
-	return k8sClient.CiliumV2().CiliumNetworkPolicies(cnp.ObjectMeta.Namespace).UpdateStatus(cnp)
+	return ciliumK8sClient.CiliumV2().CiliumNetworkPolicies(cnp.ObjectMeta.Namespace).UpdateStatus(cnp)
 }
 
 func updateOrCreateCNP(cnp *cilium_v2.CiliumNetworkPolicy) (*cilium_v2.CiliumNetworkPolicy, error) {
-	k8sClient, err := getK8sClient()
-	if err != nil {
-		return nil, err
-	}
-	k8sCNP, err := k8sClient.CiliumV2().CiliumNetworkPolicies(cnp.ObjectMeta.Namespace).
+	k8sCNP, err := ciliumK8sClient.CiliumV2().CiliumNetworkPolicies(cnp.ObjectMeta.Namespace).
 		Get(cnp.ObjectMeta.Name, v1.GetOptions{})
 	if err == nil {
 		k8sCNP.ObjectMeta.Labels = cnp.ObjectMeta.Labels
 		k8sCNP.Spec = cnp.Spec
 		k8sCNP.Specs = cnp.Specs
 		k8sCNP.Status = cilium_v2.CiliumNetworkPolicyStatus{}
-		return k8sClient.CiliumV2().CiliumNetworkPolicies(cnp.ObjectMeta.Namespace).Update(k8sCNP)
+		return ciliumK8sClient.CiliumV2().CiliumNetworkPolicies(cnp.ObjectMeta.Namespace).Update(k8sCNP)
 	}
-	return k8sClient.CiliumV2().CiliumNetworkPolicies(cnp.ObjectMeta.Namespace).Create(cnp)
+	return ciliumK8sClient.CiliumV2().CiliumNetworkPolicies(cnp.ObjectMeta.Namespace).Create(cnp)
 }
 
 func updateDerivativeStatus(cnp *cilium_v2.CiliumNetworkPolicy, derivativeName string, err error) error {
@@ -158,13 +139,9 @@ func updateDerivativeStatus(cnp *cilium_v2.CiliumNetworkPolicy, derivativeName s
 		status.OK = true
 	}
 
-	k8sClient, clientErr := getK8sClient()
-	if clientErr != nil {
-		return fmt.Errorf("Cannot get Kubernetes apiserver client: %s", clientErr)
-	}
 	// This CNP can be modified by cilium agent or operator. To be able to push
 	// the status correctly fetch the last version to avoid updates issues.
-	k8sCNPStatus, clientErr := k8sClient.CiliumV2().
+	k8sCNPStatus, clientErr := ciliumK8sClient.CiliumV2().
 		CiliumNetworkPolicies(cnp.ObjectMeta.Namespace).
 		Get(cnp.ObjectMeta.Name, v1.GetOptions{})
 	if clientErr != nil {
@@ -174,11 +151,89 @@ func updateDerivativeStatus(cnp *cilium_v2.CiliumNetworkPolicy, derivativeName s
 		// This case should not happen, but if the UID does not match make sure
 		// that the new policy is not in the cache to not loop over it. The
 		// kubernetes watcher should take care about that.
-		groupsCNPCache.DeleteCNP(k8sCNPStatus)
+		cnpCache.DeleteCNP(k8sCNPStatus)
 		return fmt.Errorf("Policy UID mistmatch")
 	}
 	k8sCNPStatus.SetDerivedPolicyStatus(derivativeName, status)
-	groupsCNPCache.UpdateCNP(k8sCNPStatus)
+	cnpCache.UpdateCNP(k8sCNPStatus)
 	_, err = updateCNPStatus(k8sCNPStatus)
+	return err
+}
+
+// DeleteDerivativeFromCache deletes the given CNP from the cnpCache to
+// no continue pooling new data.
+func DeleteDerivativeFromCache(cnp *cilium_v2.CiliumNetworkPolicy) {
+	cnpCache.DeleteCNP(cnp)
+}
+
+// DeleteDerivativeCNP if the given policy has a derivative constraint,the
+// given CNP will be deleted from store and the cache.
+func DeleteDerivativeCNP(cnp *cilium_v2.CiliumNetworkPolicy) error {
+
+	scopedLog := log.WithFields(logrus.Fields{
+		logfields.CiliumNetworkPolicyName: cnp.ObjectMeta.Name,
+		logfields.K8sNamespace:            cnp.ObjectMeta.Namespace,
+	})
+
+	if !cnp.RequiresDerivative() {
+		scopedLog.Debug("CNP does not have derivative policies, skipped")
+		return nil
+	}
+
+	err := ciliumK8sClient.CiliumV2().CiliumNetworkPolicies(cnp.ObjectMeta.Namespace).DeleteCollection(
+		&v1.DeleteOptions{},
+		v1.ListOptions{LabelSelector: fmt.Sprintf("%s=%s", parentCNP, cnp.ObjectMeta.UID)})
+	if err != nil {
+		return err
+	}
+
+	DeleteDerivativeFromCache(cnp)
+	return nil
+}
+
+func addDerivativeCNP(cnp *cilium_v2.CiliumNetworkPolicy) error {
+
+	scopedLog := log.WithFields(logrus.Fields{
+		logfields.CiliumNetworkPolicyName: cnp.ObjectMeta.Name,
+		logfields.K8sNamespace:            cnp.ObjectMeta.Namespace,
+	})
+
+	var derivativeCNP *cilium_v2.CiliumNetworkPolicy
+	var derivativeErr error
+
+	// The maxNumberOfAttempts is to not hit the limits of cloud providers API.
+	// Also, the derivativeErr is never returned, if not the controller will
+	// hit this function and the cloud providers limit will be raised. This
+	// will cause a disaster, due all other policies will hit the limit as
+	// well.
+	// If the createDerivativeCNP() fails, a new all block rule will be inserted and
+	// the derivative status in the parent policy  will be updated with the
+	// error.
+	for numAttempts := 0; numAttempts <= maxNumberOfAttempts; numAttempts++ {
+		derivativeCNP, derivativeErr = createDerivativeCNP(cnp)
+		if derivativeErr == nil {
+			break
+		}
+		scopedLog.WithError(derivativeErr).Error("Cannot create derivative rule. Installing deny-all rule.")
+		statusErr := updateDerivativeStatus(cnp, derivativeCNP.ObjectMeta.Name, derivativeErr)
+		if statusErr != nil {
+			scopedLog.WithError(statusErr).Error("Cannot update CNP status for derivative policy")
+		}
+		time.Sleep(sleepDuration)
+	}
+	cnpCache.UpdateCNP(cnp)
+	_, err := updateOrCreateCNP(derivativeCNP)
+	if err != nil {
+		statusErr := updateDerivativeStatus(cnp, derivativeCNP.ObjectMeta.Name, err)
+		if statusErr != nil {
+			scopedLog.WithError(err).Error("Cannot update CNP status for derivative policy")
+		}
+		return statusErr
+	}
+
+	err = updateDerivativeStatus(cnp, derivativeCNP.ObjectMeta.Name, nil)
+	if err != nil {
+		scopedLog.WithError(err).Error("Cannot update CNP status for derivative policy")
+	}
 	return err
 }
